# contracts/openapi/1.0.0/common.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "INSIGHT Common Module Interface",
    "version": "1.0.0",
    "description": "Shared transport routes and reusable interface schemas. Domain behavior remains module-owned."
  },
  "paths": {
...
    "/schemas/{version}/{name}": {
      "get": {
        "operationId": "getVersionedSchema",
        "parameters": [
          {"name": "version", "in": "path", "required": true, "schema": {"$ref": "../../schemas/1.0.0/identifiers.schema.json#/$defs/SemVer"}},
          {"name": "name", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-z][a-z0-9-]{0,63}$"}}
        ],
        "responses": {
          "200": {"description": "Versioned JSON Schema", "content": {"application/schema+json": {"schema": {"type": "object"}}}},
          "404": {"$ref": "#/components/responses/Problem"}
        }
      }
    }
  },
  "components": {
...
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem details",
        "content": {"application/problem+json": {"schema": {"$ref": "../../schemas/1.0.0/problem-details.schema.json"}}}
      }
    }
  }
}